+++ pkgs/code_builder/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Correct type annotations on nullable and generic variables created with
   `declareVar`, `declareFinal`, and `declareConst`.
+* Deprecate and remove uses of `Expression.expression`.
 
 ## 4.11.1